+++ fakeredis/_tcp_server.py
@@ -60,8 +60,12 @@ class Writer:
     def write(self, value: bytes) -> None:
         LOGGER.debug(f"<<< {self.client_address}: {value}")
         self.writer.write(value)
-        self.writer.flush()
 
+    def dump(self, value: Any, dump_bulk: bool = False) -> None:
+        raise NotImplementedError
+
+
+class Resp2Writer(Writer):
     def dump(self, value: Any, dump_bulk: bool = False) -> None:
         if isinstance(value, int):
             self.write(f":{value}\r\n".encode())
@@ -85,6 +89,51 @@ class Writer:
             else:
                 prefix = _get_exception_prefix(value)
                 self.write(f"-{prefix} {value.args[0]}\r\n".encode())
+        self.writer.flush()
+
+
+class Resp3Writer(Writer):
+    def dump(self, value: Any, dump_bulk: bool = False) -> None:
+        value_type = type(value)
+        if value is None:
+            self.write("_\r\n".encode())
+        elif value_type is str or value_type is bytes:
+            value = to_bytes(value)
+            if value.upper() == b"SHUTDOWN":
+                self.request_handler.shutdown_request = True
+            if dump_bulk or b"\r" in value or b"\n" in value:
+                self.write(b"$" + str(len(value)).encode() + b"\r\n" + value + b"\r\n")
+            else:
+                self.write(b"+" + value + b"\r\n")
+        elif value_type is int:
+            if -(2**63) <= value <= 2**63 - 1:  # regular integer
+                self.write(f":{value}\r\n".encode())
+            else:  # big integer
+                self.write(f"({value}\r\n".encode())
+        elif value_type is float:
+            self.write(f",{value:.17g}\r\n".encode())
+        elif value_type is list:
+            self.write(f"*{len(value)}\r\n".encode())
+            for item in value:
+                self.dump(item, dump_bulk=True)
+        elif value_type is set:
+            self.write(f"~{len(value)}\r\n".encode())
+            for item in value:
+                self.dump(item, dump_bulk=True)
+        elif value_type is bool:
+            self.write(f"#{'t' if value else 'f'}\r\n".encode())
+        elif value_type is dict:
+            self.write(f"%{len(value)}\r\n".encode())
+            for k, v in value.items():
+                self.dump(k, dump_bulk=True)
+                self.dump(v, dump_bulk=True)
+        elif isinstance(value, Exception):
+            if isinstance(value, SimpleError):
+                self.write(f"-{value.args[0]}\r\n".encode())
+            else:
+                prefix = _get_exception_prefix(value)
+                self.write(f"-{prefix} {value.args[0]}\r\n".encode())
+        self.writer.flush()
 
 
 class TCPFakeRequestHandler(StreamRequestHandler):
@@ -100,7 +149,7 @@ class TCPFakeRequestHandler(StreamRequestHandler):
         if self.client_address in self.server.clients:
             self.current_client = self.server.clients[self.client_address]
         else:
-            self.writer = Writer(self.client_address, self.wfile, self)
+            self.writer = Resp3Writer(self.client_address, self.wfile, self)
             self.current_client = FakeConnection(
                 server=self.server.fake_server,
                 writer=self.writer,