+++ fakeredis/model/_base_type.py
@@ -1,0 +1,6 @@
+class BaseModel:
+    _model_type = None
+
+    @classmethod
+    def model_type(cls) -> bytes:
+        return cls._model_type